Move `Map` instance into `map` service
